Split reductions into multiple source files
